Round proxy fee up
